Update Order,request,response

See merge request hcm25_cpl_net_06/team03_be!57

+++ WebAPI/Domain/Entities/SeatTypePrice.cs
@@ -11,6 +11,6 @@ namespace Domain.Entities
     {
         public SeatTypes SeatType { get; set; } // Enum
 
-        public double DefaultPrice { get; set; }
+        public decimal DefaultPrice { get; set; }
     }
 }
